+++ src/old_bin/book1-chapter4.rs
@@ -11,7 +11,7 @@ pub fn ray_color(ray: &Ray) -> Color {
 fn main() {
     let aspect_ratio = 16. / 9.;
     let image_width: usize = 400;
-    let image_height = (image_width as f64 / aspect_ratio) as usize;
+    let image_height = (image_width as f32 / aspect_ratio) as usize;
 
     let viewport_height = 2.0;
     let viewport_width = aspect_ratio * viewport_height;
@@ -26,8 +26,8 @@ fn main() {
     let mut ppm = Vec::<Color>::new();
     for j in (0..image_height).rev() {
         for i in 0..image_width {
-            let u = i as f64 / (image_width - 1) as f64;
-            let v = j as f64 / (image_height - 1) as f64;
+            let u = i as f32 / (image_width - 1) as f32;
+            let v = j as f32 / (image_height - 1) as f32;
             let ray = Ray::new(
                 origin,
                 lower_left_corner + u * horizontal + v * vertical - origin,
