+++ src/shapes.rs
@@ -11,10 +11,10 @@ impl Position for Stationary {}
 
 pub struct Moving {
     pub position: (Point3, Point3),
-    pub time: (f64, f64),
+    pub time: (f32, f32),
 }
 impl Moving {
-    pub fn position(&self, time: f64) -> Point3 {
+    pub fn position(&self, time: f32) -> Point3 {
         self.position.0
             + ((time - self.time.0) / (self.time.1 - self.time.0))
                 * (self.position.1 - self.position.0)
@@ -26,12 +26,12 @@ type MaterialArc = Arc<dyn Material>;
 
 pub struct Sphere<P: Position> {
     center: P,
-    radius: f64,
+    radius: f32,
     material: MaterialArc,
 }
 
 impl Sphere<Stationary> {
-    pub fn new(center: Point3, radius: f64, material: MaterialArc) -> Self {
+    pub fn new(center: Point3, radius: f32, material: MaterialArc) -> Self {
         Self {
             center: Stationary { position: center },
             radius,
@@ -39,7 +39,7 @@ impl Sphere<Stationary> {
         }
     }
 
-    pub fn with_time(self, position_end: Point3, time_start: f64, time_end: f64) -> Sphere<Moving> {
+    pub fn with_time(self, position_end: Point3, time_start: f32, time_end: f32) -> Sphere<Moving> {
         Sphere::<Moving> {
             center: Moving {
                 position: (self.center.position, position_end),
@@ -52,7 +52,7 @@ impl Sphere<Stationary> {
 }
 
 impl Hittable for Sphere<Stationary> {
-    fn hit(&self, ray: Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = ray.origin() - self.center.position;
         let a = ray.direction().norm_sq();
         let b_halves = oc.dot(&ray.direction());
@@ -83,7 +83,7 @@ impl Hittable for Sphere<Stationary> {
 }
 
 impl Hittable for Sphere<Moving> {
-    fn hit(&self, ray: Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = ray.origin() - self.center.position(ray.time());
         let a = ray.direction().norm_sq();
         let b_halves = oc.dot(&ray.direction());
@@ -115,13 +115,13 @@ impl Hittable for Sphere<Moving> {
 
 pub struct Cylinder {
     center: Point3,
-    radius: f64,
-    height: f64,
+    radius: f32,
+    height: f32,
     material: MaterialArc,
 }
 
 impl Cylinder {
-    pub fn new(center: Point3, radius: f64, height: f64, material: MaterialArc) -> Self {
+    pub fn new(center: Point3, radius: f32, height: f32, material: MaterialArc) -> Self {
         Self {
             center,
             radius,
@@ -132,7 +132,7 @@ impl Cylinder {
 }
 
 impl Hittable for Cylinder {
-    fn hit(&self, ray: Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = Point3::new(
             ray.origin().x() - self.center.x(),
             0.,
@@ -160,7 +160,7 @@ impl Hittable for Cylinder {
         let lower_bound = self.center.y() - self.height / 2.;
 
         let mut point: Point3;
-        let mut root: f64;
+        let mut root: f32;
 
         if point1.y() > upper_bound {
             if point2.y() > upper_bound {
